Refactor standalone tests to use shared mock library
- Create shared mock_gl_standalone and mock_stb_image_standalone.
- Refactor test_texture_toctou.c, test_billboard_cleanup.c, and test_shader_path_security_standalone.c to remove inline mocks.
- Update tests/CMakeLists.txt to link against shared mocks.
- Ensure all standalone tests use Unity framework consistently.
